+++ src/imio/smartweb/core/browser/sitemap.py
@@ -1,8 +1,5 @@
 from Acquisition import aq_inner
 from BTrees.OOBTree import OOBTree
-from imio.smartweb.core.config import DIRECTORY_URL
-from imio.smartweb.core.config import EVENTS_URL
-from imio.smartweb.core.config import NEWS_URL
 from imio.smartweb.core.contents.rest.directory.endpoint import DirectoryEndpointGet
 from imio.smartweb.core.contents.rest.events.endpoint import EventsEndpointGet
 from imio.smartweb.core.contents.rest.news.endpoint import NewsEndpointGet
@@ -10,6 +7,7 @@ from imio.smartweb.core.interfaces import IImioSmartwebCoreLayer
 from plone.app.layout.navigation.navtree import buildFolderTree
 from plone.app.layout.sitemap.sitemap import SiteMapView
 from plone.base.interfaces import IPloneSiteRoot
+from plone.memoize import ram
 from Products.CMFPlone.browser.navigation import CatalogSiteMap as BaseCatalogSiteMap
 from Products.CMFPlone.browser.navtree import (
     SitemapNavtreeStrategy as BaseSitemapNavtreeStrategy,
@@ -25,6 +23,7 @@ from zope.interface import Interface
 
 import logging
 import Missing
+import time
 
 logger = logging.getLogger("imio.smartweb.core")
 
@@ -44,14 +43,18 @@ FRIENDLY_TYPES = [
 ]
 
 
+def cache_key(method, obj, request):
+    """We cache data from authentic sources for the sitemap (.xml.gz) for 2 hours."""
+    return f"sitemap_{obj.UID()}_{time.time() // 7200}"
+
+
+@ram.cache(cache_key)
 def get_endpoint_data(obj, request):
-    """Retourne les données issues du bon endpoint pour un objet donné"""
     endpoint_mapping = {
         "imio.smartweb.DirectoryView": DirectoryEndpointGet,
         "imio.smartweb.EventsView": EventsEndpointGet,
         "imio.smartweb.NewsView": NewsEndpointGet,
     }
-
     endpoint_class = endpoint_mapping.get(obj.portal_type)
     if not endpoint_class:
         return {}
@@ -67,7 +70,7 @@ def get_endpoint_data(obj, request):
 
 
 def format_sitemap_items(items, base_url):
-    """Formatte les items pour le sitemap"""
+    """Format items for sitemap(.xml.gz)"""
     formatted_items = []
     for item in items:
         item_id = normalizeString(item.get("title"))
@@ -77,14 +80,14 @@ def format_sitemap_items(items, base_url):
             {
                 "loc": f"{base_url}/{item_id}?u={item_uid}",
                 "lastmod": lastmod,
-                "Title": item_id,
+                "Title": item.get("title"),
                 "Description": item.get("description", ""),
                 "getURL": f"{base_url}/{item_id}?u={item_uid}",
                 "getRemoteUrl": Missing.Value,
                 "currentItem": False,
                 "currentParent": False,
                 "normalized_review_state": "published",
-                "normalized_portal_type": "imio-smartweb-directory-item",
+                "normalized_portal_type": "imio-smartweb-authsources-item",
             }
         )
     return formatted_items
@@ -132,13 +135,14 @@ class CustomSiteMapView(SiteMapView):
             )
             yield {"loc": loc, "lastmod": modified[1]}
 
-        for brain in catalog(
+        brains = catalog(
             portal_type=[
                 "imio.smartweb.EventsView",
                 "imio.smartweb.NewsView",
                 "imio.smartweb.DirectoryView",
             ]
-        ):
+        )
+        for brain in brains:
             obj = brain.getObject()
             data = get_endpoint_data(obj, obj.REQUEST)
             yield from format_sitemap_items(data.get("items", {}), obj.absolute_url())
@@ -169,7 +173,7 @@ class CatalogSiteMap(BaseCatalogSiteMap):
 
 
 class SitemapQueryBuilder(BaseSitemapQueryBuilder):
-    """Construit la requête pour le sitemap avec des types de contenu personnalisés."""
+    """Builds the request for the sitemap with custom content types."""
 
     def __init__(self, context):
         self.context = context